requirements.txt
streamlit
spacy
newspaper3k
en_core_web_sm